backend/requirements.txt
uvicorn==0.34.0 # библиотека для запуска ASGI сервера
nest_asyncio==1.6.0 # библиотека для работы с asyncio
fastapi==0.115.8 # библиотека для создания веб-приложений
SQLAlchemy==1.4.51 # библиотека для работы с базами данных
passlib==1.7.4 # библиотека для хеширования паролей
pyjwt==2.3.0 # библиотека для работы с JWT
python-multipart==0.0.20 # библиотека для работы с multipart/form-data
paho-mqtt==1.5.1 # библиотека для работы с MQTT
crontab==0.23.0 # библиотека для работы с cron
future==1.0.0 # библиотека для работы с Python 2